Schema.sql fix
I fixed a couple of items within the schema.sql. First I DROP TABLE IF EXIST the each table to present them. The i removed unecessary NOT NULL from some of the table requirements. If an item is auot incremented, there is no need to add NOT NULL because theres not input. Also I removed the FORERIGN KEY and referencial data as this establishes basic structure while the inquire prompts will add them data themselves.