Server.js Prompts
I usually break this much code in separate commits but I got on a role of actually figuring this stuff out so I just kept coding lol. SO far I have completed the start menu that lets you decide what action to want to take. THen I have all the viewing of departments, roles, and employees that pulls from the seeds.sql. NOw Im working of the addition of new departments/etc. Im not sure if I am doing this correctly but I believe i have the inquirer.prompts right. The response in the connection.query is troublesome to me because I dont know if I have the correct syntax for the SQL commands and adding the answers to the tables. Also Im not sure how to get the seeds.sql data to present like a list while in the inquirer prompt for the user to select.