addRole function and npm init
I believe Ive created a functioning inquirer prompt that pulls from the seeds.sql file to display the departments available in the company db. I tried to do some trouble shooting but nothing was working. THtas when I figured out that I forgot to do npm init and none of the installs that I had were present. I always forget to do that.