addRole department ids and strings
I finally have a working addRole function. I made a inquirer prompt list with all the departments from the seeds.sql file by using a connect.query to initiate a sql command and created a departmentNames object to call in the choices section of the prompt. However I had an issue with the list data being strings and the data needed was integers based on department_ids. So i used a reduce function that takes in an accumulator argument and a currency argument to set the property of the entery data to be a integer based on the string choice the user makes.