addEmployee
Ive added my addEmployee function which I modeled after my addRole function. Its essentially the same premise but using two connection queries with different syntax to pull the role list and amanger list. Also I was careful to add the appropriate accumulative and currency code to ensure that the value being accepted is a string but the value being processed by the list is an integer because of the ids.