addEmployee Troubleshoot
There was several rpoblems with the code that I fixed. First the name for the roles needed to be changed to title because thats what it is defined as in the seeds.sql file. Also in order to show every employee (not just ones that arent current managers) I needed to change the connection.query to SELECT the manager_id and use a forEach method to take in all the employees.