View all employees function and update employees
After finishing the update employees function I found that the view all employees function doesnt meet the challenges requirements so I change the code to use the connection.query and new syntax (thanks stackoverflow) to choose each item that needs to be presented in the terminal. The e, r, d are all references to the seeds.sql tables employee, role, department. I can present each item in order by defining which table I want to pull from and the data to present from that table.