Delete Functions
I added for extra credit the delete fucntions for department, employee, and role. These are generally straight forward in that it follows the same structure and syntax as the add functions except that you use a connection.query to delete based on id. Also I included a list inquirer prompt so you can see who is being deleted as well as a confirmation prompt to prevent mistakes.